Update logout onclick function

+++ client/src/components/Messenger/Sidebar/Dropdown.jsx
@@ -1,9 +1,13 @@
 import styles from './Dropdown.module.css';
-import { useState } from 'react';
+import { useContext, useState } from 'react';
 import { Navigate } from 'react-router-dom';
+import { chatContext } from '../../../contexts/chatContext';
+import LogoutFetch from '../../../fetch/LogoutAPI';
 
 const Dropdown = ({ setShowUserList, setShowFriendList, setShowSettings }) => {
+  const { setError } = useContext(chatContext);
   const [appRedirect, setAppRedirect] = useState(false);
+  const [loading, setLoading] = useState(false);
 
   const onSettings = (e) => {
     e.stopPropagation();
@@ -12,23 +16,42 @@ const Dropdown = ({ setShowUserList, setShowFriendList, setShowSettings }) => {
     setShowFriendList(false);
   };
 
-  const onLogOut = (e) => {
+  const onLogOut = async (e) => {
     e.stopPropagation();
-    localStorage.clear();
-    setAppRedirect(true);
+    setLoading(true);
+
+    const result = await LogoutFetch();
+
+    if (result && result.error) {
+      setError(true);
+    }
+
+    if (result && result.responseData) {
+      localStorage.clear();
+      setLoading(false);
+      setAppRedirect(true);
+    }
   };
 
   return (
-    <div className={styles.Dropdown} id="dropdown" data-testid="dropdown">
-      <div className={styles.button} onClick={onSettings}>
-        <button>Settings</button>
-      </div>
-      <hr id="dropdown"></hr>
-      <div className={styles.button} onClick={onLogOut}>
-        <button>Log Out</button>
-      </div>
-      {appRedirect && <Navigate to="/" replace={true} />}
-    </div>
+    <>
+      {loading ? (
+        <div className={styles.loading} data-testid="loading">
+          <div className={styles.loader}></div>
+        </div>
+      ) : (
+        <div className={styles.Dropdown} id="dropdown" data-testid="dropdown">
+          <div className={styles.button} onClick={onSettings}>
+            <button>Settings</button>
+          </div>
+          <hr id="dropdown"></hr>
+          <div className={styles.button} id="logout" onClick={onLogOut}>
+            <button id="logout">Log Out</button>
+          </div>
+          {appRedirect && <Navigate to="/" replace={true} />}
+        </div>
+      )}
+    </>
   );
 };
